requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-allauth
requests
cryptography